requirements.txt
apscheduler==3.10.4
psutil==5.9.5
PyDispatcher==2.0.7
pyserial==3.5
Requests==2.31.0
